# Optimize job dispatch performance