--- a/.gitignore
+++ b/.gitignore
@@ -90,3 +90,4 @@ node_modules
 
 docs/.vuepress/dist
 /node_modules
+BenchmarkDotNet.Artifacts