--- a/src/Quartz.Benchmark/JobDispatchBenchmark.cs
+++ b/src/Quartz.Benchmark/JobDispatchBenchmark.cs
@@ -1,0 +1,37 @@
+using System;
+using System.Threading.Tasks;
+
+using BenchmarkDotNet.Attributes;
+
+using Quartz.Core;
+using Quartz.Impl;
+using Quartz.Job;
+using Quartz.Spi;
+
+namespace Quartz.Benchmark
+{
+    [MemoryDiagnoser]
+    public class JobDispatchBenchmark
+    {
+        private StdScheduler scheduler;
+        private JobRunShell shell;
+
+        [GlobalSetup]
+        public void Setup()
+        {
+            scheduler = (StdScheduler) new StdSchedulerFactory().GetScheduler().GetAwaiter().GetResult();
+            var job = JobBuilder.Create<NoOpJob>().Build();
+            var trigger = (IOperableTrigger) TriggerBuilder.Create().ForJob(job.Key).WithSimpleSchedule().StartNow().Build();
+            trigger.FireInstanceId = "fire-instance-id";
+            var bundle = new TriggerFiredBundle(job, trigger, null, false, DateTimeOffset.UtcNow, null, null, null);
+            shell = new JobRunShell(scheduler, bundle);
+        }
+        
+        [Benchmark]
+        public async Task Run()
+        {
+            await shell.Initialize(scheduler.sched);
+            await shell.Run();
+        }
+    }
+}