--- a/src/Quartz.Benchmark/Program.cs
+++ b/src/Quartz.Benchmark/Program.cs
@@ -1,0 +1,34 @@
+﻿using System.Runtime.CompilerServices;
+
+using BenchmarkDotNet.Running;
+
+namespace Quartz.Benchmark
+{
+    internal class Program
+    {
+        private static void Main(string[] args)
+        {
+            BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run();
+
+            //DispatchBenchmark();
+        }
+
+        private static void DispatchBenchmark()
+        {
+            var benchmark = new JobDispatchBenchmark();
+            benchmark.Setup();
+            benchmark.Run().GetAwaiter().GetResult();
+
+            RunDispatch(benchmark);
+        }
+
+        [MethodImpl(MethodImplOptions.NoInlining)]
+        private static void RunDispatch(JobDispatchBenchmark benchmark)
+        {
+            for (int i = 0; i < 100; ++i)
+            {
+                benchmark.Run().GetAwaiter().GetResult();
+            }
+        }
+    }
+}