--- a/src/Quartz.Examples/01_SimpleJobScheduler/HelloJob.cs
+++ b/src/Quartz.Examples/01_SimpleJobScheduler/HelloJob.cs
@@ -39,7 +39,8 @@ namespace Quartz.Examples.Example01
         public virtual Task Execute(IJobExecutionContext context)
         {
             // Say Hello to the World and display the date/time
-            Console.WriteLine($"Hello World! - {DateTime.Now:r}");
+            var timestamp = DateTime.Now;
+            Console.WriteLine($"Hello World! - {timestamp:yyyy-MM-dd HH:mm:ss.fff}");
             return Task.CompletedTask;
         }
     }