… changelog.md
… src/Quartz/Collections/OrderedDictionary.ValueCollection.cs
… src/Quartz/Collections/OrderedDictionary.cs
… src/Quartz/ExceptionHelper.cs
… src/Quartz/Impl/JobExecutionContextImpl.cs
… src/Quartz/JobDataMap.cs
… src/Quartz/Simpl/PropertySettingJobFactory.cs
… src/Quartz/TaskExtensions.cs
… src/Quartz/Util/ObjectUtils.cs